test ArgumentTypeError
